requirements.txt
Django==4.2.2
django-filter==23.2
djangorestframework==3.14.0
drf-yasg==1.21.5
psycopg2==2.9.6
coverage==7.2.7
